refactor: extract saveVmConnection to history.ts
Eliminates copy-paste of saveVmConnection across 6 cloud provider files.
Fixes #1787

Agent: complexity-hunter
