+++ cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@openrouter/spawn",
-  "version": "0.7.4",
+  "version": "0.7.5",
   "type": "module",
   "bin": {
     "spawn": "cli.js"
